// precon-platform-ui/src/types/ProjectNormals.ts
export default interface ProjectNormals {
  id: number;
  month: number;
  avgHighTemp?: number;
  avgLowTemp?: number;
  recordAvgHighTemp?: number;
  recordAvgLowTemp?: number;
  avgRain?: number;
  recordHighAvgRain?: number;
  recordLowAvgRain?: number;
  avgSnow?: number;
  recordHighAvgSnow?: number;
  recordLowAvgSnow?: number;
  avgHighWind?: number;
  avgLowWind?: number;
  recordAvgHighWind?: number;
  recordAvgLowWind?: number;
}